Switch to newer arbor-go API
This change allows users of arbor-go to handle errors when reading and
writing protocol-level messages. This commit switches the server to use
the newer API. It also switches the server to the `validate` branch of
arbor-go in order to prove that the validation changes there prevent
server crashes for certain malformed messages.